코딩테스트 챌린지 Day02
코딩테스트 챌린지 Day02

+++ javaBasic/src/codingTestChallenge/Day1.java
@@ -8,8 +8,8 @@ public class Day1 {
 		
 		Scanner sc = new Scanner(System.in);		
 		long a = sc.nextLong();
-		long b = sc.nextLong();
-		long c = sc.nextLong();
+		long b = sc.nextInt();
+		long c = sc.nextInt();
 		
 		System.out.println(a+b+c);
